Refactor Week class to improve clarity and coherence
The Week class was refactored with various changes to improve clarity and coherence of the code. The term 'entrys' was changed to 'lession' to better reflect its use and the 'config.php' was replaced with 'config.local.php' for local development purposes. The change from 'config.php' to 'config.local.php' was also added to .gitignore to prevent environment specific configurations from being committed. PHPDoc comments were added to methods for clear understanding of their functionality. The variable $modul's data type was fixed to string to ensure type integrity.