"Dynamically position checkboxes in oneweek.css view"
The positioning of checkboxes in oneweek.css has been made dynamic. The top position of each checkbox, previously hard-coded, has been removed in favor of a dynamic one. This change was propagated to oneweek.php, where corresponding increases in top-position are dynamically handled in echo statements. This helps the UI to be flexible, especially when the list of options changes.